+++ ixmp/backend/base.py
@@ -842,20 +842,31 @@ class Backend(ABC):
     @abstractmethod
     def export_timeseries_data(self, file, default, model, scenario,
                                variables):
-        """Export timeseries data to file
+        """Export timeseries data to CSV file
 
         Parameters
         ----------
         file : str
-            File name to export data to
+            File name to export data to.
+            Result file will contain following columns:
+            - model
+            - scenario
+            - version
+            - variable
+            - unit
+            - region
+            - meta
+            - time
+            - year
+            - value
         default : bool
-           :obj:`True` to include only TimeSeries versions marked as default.
+            :obj:`True` to include only TimeSeries versions marked as default.
         model : str or None
-           Model name to filter results.
+            Model name to filter results.
         scenario : str or None
-           Scenario name to filter results.
+            Scenario name to filter results.
         variables : list
-            List of timeseries variables to export
+            List of timeseries variables (names) to export
 
         Returns
         -------